fix(path): Remove non-node symbols
